chore: organize page component location
