+++ src/components/Layout.tsx
@@ -1,5 +1,5 @@
 import { Slot, component$ } from '@builder.io/qwik';
-import { Route } from '../utils/router';
+import { Route } from '../router';
 import { Header } from './Header';
 
 export const Layout = component$<{ currentRoute: Exclude<Route, 'auth'> }>(({ currentRoute }) => {
