+++ src/components/SkillMatrix.tsx
@@ -1,10 +1,10 @@
 import { $, component$, useContext, useSignal, useTask$ } from '@builder.io/qwik';
 import { AppContext } from '../app';
 import { tt } from '../locale/labels';
+import { navigateTo } from '../router';
 import { getSkillMatrixMine, pathSkillMatrixMine } from '../utils/api';
 import { COOKIE_TOKEN_KEY } from '../utils/constants';
 import { getCookie } from '../utils/cookie';
-import { navigateTo } from '../utils/router';
 import { Skill } from '../utils/types';
 import { SkillRow } from './SkillRow';
